+++ scripts/version/config.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+#include <json-c/json.h>
+
+int main(void)
+{
+	return 0;
+}